MUU-357: changed field 338, subfield b.
