+++ test-fixtures/muuntaja/e2print/default/f020_776/input.json
@@ -1,6 +1,6 @@
 {
  "options": {
-  "type": "p2e",
+  "type": "e2p",
   "profile": "KVP",
   "format": "PDF",
   "LOWTAG":"KVP"
