+++ test-fixtures/muuntaja/e2print/fennica/f020_776/input.json
@@ -1,8 +1,8 @@
 {
  "options": {
-  "type": "p2e",
-  "profile": "KVP",
-  "LOWTAG":"KVP"
+  "type": "e2p",
+  "profile": "FENNICA",
+  "LOWTAG":"FENNICA"
  },
  "source": {},
  "base": {},
