+++ test-fixtures/muuntaja/e2print/fennica/f338/metadata.json
@@ -1,0 +1,10 @@
+{
+  "description": "e2p - fennica - 338",
+  "skip": false,
+  "only": false,
+  "succesRecords": 2,
+  "failedRecords": 0,
+  "testBase": true,
+  "expectedError": false,
+  "expectedErrorStatus": "200"
+}
