+++ test-fixtures/muuntaja/print2e/default/f338/input.json
@@ -1,0 +1,12 @@
+{
+  "options": {
+   "type": "p2e",
+   "profile": "KVP",
+   "LOWTAG":"KVP"
+  },
+  "source": {},
+  "base": {},
+  "exclude": {},
+  "replace": {},
+  "include": []
+ }
