+++ test-fixtures/muuntaja/print2e/fennica/f020_776/input.json
@@ -1,7 +1,7 @@
 {
  "options": {
   "type": "p2e",
-  "profile": "KVP",
+  "profile": "FENNI",
   "format": "PDF",
   "LOWTAG":"KVP"
  },
