+++ test-fixtures/muuntaja/print2e/fennica/f338/input.json
@@ -1,0 +1,12 @@
+{
+  "options": {
+   "type": "p2e",
+   "profile": "FENNICA",
+   "LOWTAG":"FENNICA"
+  },
+  "source": {},
+  "base": {},
+  "exclude": {},
+  "replace": {},
+  "include": []
+ }
