+++ test-fixtures/muuntaja/print2e/fennica/f338/metadata.json
@@ -1,0 +1,10 @@
+{
+  "description": "print2e - fennica - 338",
+  "skip": false,
+  "only": false,
+  "succesRecords": 2,
+  "failedRecords": 0,
+  "testBase": true,
+  "expectedError": false,
+  "expectedErrorStatus": "200"
+}
